Move custom strategies into main module and update imports accordingly
